server/requirements.txt
certifi==2022.12.7
cffi==1.15.1
Deprecated==1.2.13
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.4.2
gcloud==0.18.3
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==4.0.0
requests-toolbelt==0.10.1
rsa==4.9
schedule==1.1.0
six==1.16.0
urllib3==1.26.15
wrapt==1.15.0